♻️ refactor(arch): hoist ICodingAgentDeployer port to @packmind/types
Two hexagonal violations in coding-agent:

- domain/repository/ICodingAgentDeployer.ts imported the concrete
  DefaultSkillsDeployResult from infra (domain -> infra). Move the port and its
  result types (DefaultSkillMetadata, DefaultSkillsDeployResult,
  DeployDefaultSkillsOptions) into @packmind/types as the canonical contract;
  the domain now re-exports them and infra imports them from types. This also
  drops the ICodingAgentDeployer = unknown placeholder in the registry contract.

- infra PackmindDeployer imported StandardsIndexService from application
  (infra -> application). The service is a pure index-string builder, so move it
  next to PackmindDeployer under infra/repositories/packmind/.
