+++ packages/coding-agent/src/index.ts
@@ -1,10 +1,6 @@
 export * from './domain/AgentConfiguration';
 export * from './domain/repository/ICodingAgentDeployer';
-export {
-  DefaultSkillMetadata,
-  DefaultSkillsDeployer,
-  DefaultSkillsDeployResult,
-} from './infra/repositories/defaultSkillsDeployer/DefaultSkillsDeployer';
+export { DefaultSkillsDeployer } from './infra/repositories/defaultSkillsDeployer/DefaultSkillsDeployer';
 export * from './domain/repository/ICodingAgentDeployerRegistry';
 export * from './application/services/DeployerService';
 export * from './application/services/CodingAgentServices';
